# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Education Mini Program API
  version: 0.1.0
  description: >-
    Shared API contract for the first-phase education mini program. Successful
    responses use a data envelope and failures preserve their HTTP status.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Environment-relative service root
...
security:
  - bearerAuth: []
paths:
...
  /management/staff-accounts:
    get:
      operationId: listStaffAccounts
      summary: List pre-created staff accounts
      tags: [Management]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: query
          in: query
          schema: { type: string, maxLength: 100 }
        - name: roleCode
          in: query
          schema: { $ref: '#/components/schemas/ManageableStaffRole' }
        - name: campusId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/StaffAccountStatus' }
      responses:
        '200':
          description: Paginated staff accounts with masked phone numbers.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffAccountsPageEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createStaffAccount
      summary: Create one staff account with one role and one campus
      tags: [Management]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStaffAccountRequest' }
      responses:
        '201':
          description: Staff account created or an idempotent replay returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffAccountEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /management/staff-accounts/{staffAccountId}/status:
    patch:
      operationId: updateStaffAccountStatus
      summary: Enable or disable one staff account
      tags: [Management]
      parameters:
        - name: staffAccountId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateStaffAccountStatusRequest' }
      responses:
        '200':
          description: Staff account status updated or an idempotent replay returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffAccountEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /management/staff-accounts/{staffAccountId}/unbind-wechat:
    post:
      operationId: unbindStaffAccountWechat
      summary: Remove the WeChat identity bound to one staff account
      tags: [Management]
      parameters:
        - name: staffAccountId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UnbindStaffWechatRequest' }
      responses:
        '200':
          description: WeChat identity removed or an idempotent replay returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffAccountEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /parents/me/home:
    get:
      operationId: getParentHome
      summary: Read the primary bound student's home summary
      tags:
        - Parent
      parameters:
        - $ref: '#/components/parameters/ParentStudentId'
      responses:
        '200':
          description: Parent home data derived from PostgreSQL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentHomeEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /parents/me/campuses:
    get:
      operationId: listParentCampuses
      summary: List campuses published to parents, optionally sorted by distance
      tags:
        - Parent
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: latitude
          in: query
          schema: { type: number, format: double, minimum: -90, maximum: 90 }
        - name: longitude
          in: query
          schema: { type: number, format: double, minimum: -180, maximum: 180 }
      responses:
        '200':
          description: Paginated public campus locations.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParentCampusMapPageEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /parents/me/hours:
    get:
      operationId: getParentHours
      summary: Read balances and immutable lesson-ledger entries for a bound student
      tags:
        - Parent
      parameters:
        - $ref: '#/components/parameters/ParentStudentId'
      responses:
        '200':
          description: Parent lesson balance and ledger data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentHoursEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /parents/me/leave:
    get:
      operationId: getParentLeavePage
      summary: Read bound students, eligible lessons, and leave requests
      tags:
        - Parent
      parameters:
        - $ref: '#/components/parameters/ParentStudentId'
      responses:
        '200':
          description: Parent leave page data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentLeavePageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /parents/me/leave-requests:
    post:
      operationId: submitParentLeaveRequest
      summary: Submit one leave request for a bound student and lesson
      tags:
        - Parent
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentLeaveRequest'
      responses:
        '200':
          description: Created or replayed leave request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentLeaveRecordEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /parents/me/profile:
    get:
      operationId: getParentProfile
      summary: Read the authenticated parent's primary bound student
      tags:
        - Parent
      responses:
        '200':
          description: Parent profile data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentProfileEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      operationId: updateParentProfile
      summary: Update age and home address for the authenticated parent's primary bound student
      tags:
        - Parent
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentProfileUpdateRequest'
      responses:
        '200':
          description: Updated or replayed parent profile data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentProfileEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /parents/me/updates:
    get:
      operationId: getParentUpdates
      summary: Read attendance and teacher feedback for a bound student
      tags:
        - Parent
      parameters:
        - $ref: '#/components/parameters/ParentStudentId'
      responses:
        '200':
          description: Parent-visible class updates.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentUpdatesEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by the authentication module.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Caller-generated key used by mutation endpoints to make retries safe.
        Keys contain 8-128 letters, numbers, dots, underscores, colons, or
        hyphens.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]{8,128}$'
      example: lesson-confirm-20260827-0001
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ParentStudentId:
      name: studentId
      in: query
      required: false
      description: >-
        Bound student to read. When omitted, the primary active binding is used.
        Supplying an unbound or cross-campus student is forbidden.
      schema:
        type: string
        format: uuid
  headers:
    RequestId:
      description: Correlation identifier matching the response body requestId.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
...
  responses:
...
    BadRequest:
      description: Request validation failed.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Authentication is required or the session is invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: The authenticated user lacks the required role or data scope.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: The requested state transition conflicts with current state.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ResourceNotFound:
      description: The requested resource does not exist.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            missingResource:
              value:
                code: RESOURCE_NOT_FOUND
                message: Cannot GET /resource
                requestId: 64ea9657-41f1-4baf-a3fb-d62e39054016
                details: {}
    InternalServerError:
      description: An unexpected server error occurred.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            internalError:
              value:
                code: INTERNAL_SERVER_ERROR
                message: Internal server error
                requestId: 64ea9657-41f1-4baf-a3fb-d62e39054016
                details: {}
  schemas:
...
    SuccessEnvelope:
      type: object
      description: Common wrapper for every successful response.
      required:
        - data
        - requestId
      properties:
        data:
          description: Endpoint-specific response payload.
        requestId:
          type: string
          minLength: 1
          maxLength: 128
          description: Correlation identifier returned in X-Request-Id.
      additionalProperties: false
    PaginatedSuccessEnvelope:
      type: object
      description: Common wrapper for paginated list responses.
      required:
        - data
        - meta
        - requestId
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        requestId:
          type: string
          minLength: 1
          maxLength: 128
          description: Correlation identifier returned in X-Request-Id.
      additionalProperties: false
...
    ErrorEnvelope:
      type: object
      description: Common response body for failures; the HTTP status is preserved.
      required:
        - code
        - message
        - requestId
        - details
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
          maxLength: 128
        details:
          type: object
          additionalProperties: true
      additionalProperties: false
    ErrorCode:
      type: string
      enum:
        - BAD_REQUEST
        - VALIDATION_FAILED
        - INVALID_IDEMPOTENCY_KEY
        - UNAUTHORIZED
        - FORBIDDEN
        - RESOURCE_NOT_FOUND
        - METHOD_NOT_ALLOWED
        - CONFLICT
        - STAFF_PHONE_ALREADY_EXISTS
        - STAFF_ACCOUNT_UNAVAILABLE
        - STAFF_PHONE_BINDING_REQUIRED
        - STAFF_ACCOUNT_VERSION_CONFLICT
        - WECHAT_IDENTITY_CONFLICT
        - WECHAT_PROVIDER_RESPONSE_INVALID
        - LESSON_NOT_ASSIGNED
        - LESSON_ALREADY_COMPLETED
        - LESSON_VERSION_CONFLICT
        - LESSON_REVERSAL_WINDOW_EXPIRED
        - ATTENDANCE_INCOMPLETE
        - INSUFFICIENT_LESSON_BALANCE
        - FEEDBACK_NOT_ALLOWED
        - FEEDBACK_IMAGE_INVALID
        - FEEDBACK_IMAGE_CONFLICT
        - GROUP_CAMPAIGN_POSTER_INVALID
        - CUSTOMER_SERVICE_QR_INVALID
        - IDEMPOTENCY_CONFLICT
        - EARNING_RULE_NOT_FOUND
        - EARNING_RULE_OVERLAP
        - EARNING_NOT_REVIEWABLE
        - EARNING_ALREADY_REVIEWED
        - EARNING_REVERSED
        - WITHDRAWAL_AMOUNT_INVALID
        - WITHDRAWAL_BALANCE_INSUFFICIENT
        - WITHDRAWAL_FREQUENCY_EXCEEDED
        - WITHDRAWAL_STATUS_CONFLICT
        - WITHDRAWAL_SELF_REVIEW_FORBIDDEN
        - PAYOUT_PROOF_REQUIRED
        - PAYOUT_PROOF_INVALID
        - TOO_MANY_REQUESTS
        - INTERNAL_SERVER_ERROR
    PaginationMeta:
      type: object
      description: Metadata returned by paginated list endpoints.
      required:
        - page
        - pageSize
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
...
    ManageableStaffRole:
      type: string
      enum: [TEACHER, CAMPUS_MANAGER, PARTNER, HR, FINANCE]
    DirectlyCreatableStaffRole:
      type: string
      enum: [CAMPUS_MANAGER, PARTNER, HR, FINANCE]
    StaffAccountStatus:
      type: string
      enum: [ACTIVE, DISABLED]
    StaffBindingStatus:
      type: string
      enum: [BOUND, UNBOUND]
    CreateStaffAccountRequest:
      type: object
      description: 总端创建单角色账号；人力/财务为总部账号，campusId 必须为 null；管理员/合作方必须选择校区。
      required: [displayName, phone, roleCode, campusId]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 100, pattern: '.*\S.*' }
        phone: { type: string, pattern: '^\+?861[3-9][0-9]{9}$' }
        roleCode: { $ref: '#/components/schemas/DirectlyCreatableStaffRole' }
        campusId: { type: [string, 'null'], format: uuid }
      oneOf:
        - properties:
            roleCode: { enum: [CAMPUS_MANAGER, PARTNER] }
            campusId: { type: string, format: uuid }
        - properties:
            roleCode: { enum: [HR, FINANCE] }
            campusId: { type: 'null' }
      additionalProperties: false
    UpdateStaffAccountStatusRequest:
      type: object
      required: [status, expectedVersion]
      properties:
        status: { $ref: '#/components/schemas/StaffAccountStatus' }
        expectedVersion: { type: integer, minimum: 1 }
      additionalProperties: false
    UnbindStaffWechatRequest:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, minimum: 1 }
      additionalProperties: false
    StaffAccountView:
      type: object
      required: [id, displayName, maskedPhone, roleCode, campusId, campusName, bindingStatus, status, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        maskedPhone: { type: string, pattern: '^\+86\*{4}[0-9]{4}$' }
        roleCode: { $ref: '#/components/schemas/ManageableStaffRole' }
        campusId: { type: [string, 'null'], format: uuid }
        campusName: { type: string, minLength: 1, maxLength: 200 }
        bindingStatus: { $ref: '#/components/schemas/StaffBindingStatus' }
        status: { $ref: '#/components/schemas/StaffAccountStatus' }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      additionalProperties: false
...
    ParentStudentView:
      type: object
      required:
        - id
        - name
        - age
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 100
        age:
          type: integer
          minimum: 0
          maximum: 150
      additionalProperties: false
...
    ParentHomeView:
      type: object
      required:
        - student
        - nextLesson
        - remainingUnits
        - attendanceRatePercent
      properties:
        student:
          oneOf:
            - $ref: '#/components/schemas/ParentStudentView'
            - type: 'null'
        nextLesson:
          oneOf:
            - $ref: '#/components/schemas/ParentNextLessonView'
            - type: 'null'
        remainingUnits:
          type: integer
          minimum: 0
          description: Total active main and gift balance in hundredths of one lesson.
        attendanceRatePercent:
          oneOf:
            - type: integer
              minimum: 0
              maximum: 100
            - type: 'null'
      additionalProperties: false
    ParentCampusMapView:
      type: object
      required: [id, name, address, contactPhone, latitude, longitude, distanceMeters]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 200 }
        address: { type: string, minLength: 1, maxLength: 300 }
        contactPhone: { oneOf: [{ type: string, maxLength: 30 }, { type: 'null' }] }
        latitude: { type: number, format: double, minimum: -90, maximum: 90 }
        longitude: { type: number, format: double, minimum: -180, maximum: 180 }
        distanceMeters: { oneOf: [{ type: integer, minimum: 0 }, { type: 'null' }] }
      additionalProperties: false
...
    ParentHoursView:
      type: object
      required:
        - remainingTotalUnits
        - mainBalanceUnits
        - giftBalanceUnits
        - paidAmountFen
        - validUntil
        - entries
      properties:
        mainReservedUnits: { type: integer, minimum: 0 }
        giftReservedUnits: { type: integer, minimum: 0 }
        reservedTotalUnits: { type: integer, minimum: 0 }
        mainAvailableUnits: { type: integer, minimum: 0 }
        giftAvailableUnits: { type: integer, minimum: 0 }
        availableTotalUnits: { type: integer, minimum: 0 }
        totalBalanceUnits: { type: integer, minimum: 0, description: Gross active-package balance including refund reservations. }
        remainingTotalUnits:
          type: integer
          minimum: 0
          description: Gross active-package balance. Use availableTotalUnits for spendable lessons.
        mainBalanceUnits:
          type: integer
          minimum: 0
        giftBalanceUnits:
          type: integer
          minimum: 0
        paidAmountFen:
          type: integer
          minimum: 0
          description: Recorded package amount in integer fen.
        validUntil:
          oneOf:
            - type: string
              format: date
            - type: 'null'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/ParentHoursEntryView'
      additionalProperties: false
    ParentLeaveStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
    ParentLeaveRequest:
      type: object
      required:
        - studentId
        - lessonSessionId
        - reason
      properties:
        studentId:
          type: string
          format: uuid
        lessonSessionId:
          type: string
          format: uuid
        reason:
          type: string
          minLength: 1
          maxLength: 500
      additionalProperties: false
    ParentLeaveRecordView:
      type: object
      required:
        - id
        - studentId
        - lessonSessionId
        - courseName
        - lessonStartsAt
        - reason
        - status
        - reviewerName
        - reviewedAt
        - reviewReason
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        lessonSessionId:
          type: string
          format: uuid
        courseName:
          type: string
          minLength: 1
          maxLength: 200
        lessonStartsAt:
          type: string
          format: date-time
        reason:
          type: string
          minLength: 1
          maxLength: 500
        status:
          $ref: '#/components/schemas/ParentLeaveStatus'
        reviewerName:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 100
            - type: 'null'
        reviewedAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        reviewReason:
          oneOf:
            - type: string
              maxLength: 500
            - type: 'null'
      additionalProperties: false
    ParentLeaveLessonView:
      type: object
      required:
        - id
        - title
        - startsAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        startsAt:
          type: string
          format: date-time
      additionalProperties: false
    ParentLeavePageView:
      type: object
      required:
        - students
        - lessons
        - records
        - cutoffHours
      properties:
        students:
          type: array
          items:
            $ref: '#/components/schemas/ParentStudentView'
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/ParentLeaveLessonView'
        records:
          type: array
          items:
            $ref: '#/components/schemas/ParentLeaveRecordView'
        cutoffHours:
          type: integer
          minimum: 0
      additionalProperties: false
    ParentProfileView:
      type: object
      required:
        - student
        - unreadMessageCount
      properties:
        student:
          oneOf:
            - $ref: '#/components/schemas/ParentProfileStudentView'
            - type: 'null'
        unreadMessageCount:
          type: integer
          minimum: 0
      additionalProperties: false
...
    ParentProfileUpdateRequest:
      type: object
      required:
        - age
        - homeAddress
        - expectedVersion
      properties:
        age:
          type: integer
          minimum: 0
          maximum: 120
        homeAddress:
          type: string
          maxLength: 300
        expectedVersion:
          type: integer
          minimum: 1
      additionalProperties: false
...
    ParentUpdatesView:
      type: object
      required:
        - records
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/ParentUpdateRecordView'
      additionalProperties: false
...
    StaffAccountsPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/PaginatedSuccessEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/StaffAccountView' }
    StaffAccountEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/StaffAccountView' }
...
    ParentHomeEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ParentHomeView'
    ParentCampusMapPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/PaginatedSuccessEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ParentCampusMapView' }
    ParentHoursEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ParentHoursView'
    ParentLeavePageEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ParentLeavePageView'
    ParentLeaveRecordEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ParentLeaveRecordView'
    ParentProfileEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ParentProfileView'
    ParentUpdatesEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ParentUpdatesView'